fix: no initials are visible when on dialog element
Fix a problem where user name initials are not visible when the
component is inside a <dialog> element.
Due to how dialog element toggling works, getBoundingClientRect() would get
the width even before the component is displayed, so it returned 0.
Now the default fallback font-size is set to 16px when width is a falsy
value.